🧹 [code health] Extract duplicate ParticleBurst to shared component
