+++ src/games/RockPaperScissors.jsx
@@ -1,3 +1,4 @@
+import ParticleBurst from '../components/ParticleBurst'
 import { useState, useEffect, useRef, useCallback } from 'react'
 
 const OPTIONS = [
@@ -12,49 +13,6 @@ const RESULT_CONFIG = {
   Player: { label: '🎉 You Win!',   gradient: 'linear-gradient(135deg, #4ade80, #06b6d4)', particleColor: '#4ade80' },
   CPU:    { label: '💀 CPU Wins!',  gradient: 'linear-gradient(135deg, #f87171, #f59e0b)', particleColor: '#f87171' },
   Draw:   { label: "🤝 It's a Draw!", gradient: 'linear-gradient(135deg, #8b5cf6, #06b6d4)', particleColor: '#a78bfa' },
-}
-
-// ── Particle burst ──────────────────────────────────────────────────────────
-function Particle({ x, y, color, onDone }) {
-  const angle  = Math.random() * 360
-  const speed  = 60 + Math.random() * 80
-  const dx     = Math.cos((angle * Math.PI) / 180) * speed
-  const dy     = Math.sin((angle * Math.PI) / 180) * speed
-  const size   = 6 + Math.random() * 8
-  const shapes = ['●', '★', '◆', '▲', '♥']
-  const shape  = shapes[Math.floor(Math.random() * shapes.length)]
-
-  return (
-    <span
-      style={{
-        position: 'fixed',
-        left: x,
-        top: y,
-        fontSize: size,
-        color,
-        pointerEvents: 'none',
-        zIndex: 9999,
-        animation: `rps-particle 0.9s ease-out forwards`,
-        '--dx': `${dx}px`,
-        '--dy': `${dy}px`,
-      }}
-    >
-      {shape}
-    </span>
-  )
-}
-
-function ParticleBurst({ origin, color, id }) {
-  const [particles] = useState(() =>
-    Array.from({ length: 22 }, (_, i) => i)
-  )
-  return (
-    <>
-      {particles.map(i => (
-        <Particle key={`${id}-${i}`} x={origin.x} y={origin.y} color={color} />
-      ))}
-    </>
-  )
 }
 
 // ── Choice card ─────────────────────────────────────────────────────────────
@@ -223,7 +181,7 @@ export default function RockPaperScissors() {
       <style>{RPS_STYLES}</style>
 
       {/* Particle bursts */}
-      {bursts.map(b => <ParticleBurst key={b.id} id={b.id} origin={{ x: b.x, y: b.y }} color={b.color} />)}
+      {bursts.map(b => <ParticleBurst key={b.id} id={b.id} x={b.x} y={b.y} color={b.color} />)}
 
       <div className="rps-root">
         {/* Animated background orbs */}
@@ -319,10 +277,6 @@ function ScoreBox({ label, value, color }) {
 
 // ── Scoped styles ────────────────────────────────────────────────────────────
 const RPS_STYLES = `
-  @keyframes rps-particle {
-    0%   { transform: translate(0, 0) scale(1); opacity: 1; }
-    100% { transform: translate(var(--dx), var(--dy)) scale(0); opacity: 0; }
-  }
   @keyframes rps-bounce {
     0%, 100% { transform: translateY(0) scale(1); }
     40%       { transform: translateY(-18px) scale(1.15); }
